@@ -10,11 +10,11 @@
   "author": "Lola <lola@zeroid.bit>",
   "license": "WTFPL",
   "dependencies": {
+    "caniuse-lite": "^1.0.30001008",
     "mobx": "^3.2.1",
     "mobx-react": "^4.2.2",
-    "react": "^15.6.1",
+    "react": "^15.6.2",
     "react-dom": "^15.6.1",
-    "render-media": "^2.10.0",
     "semantic-ui-css": "^2.4.0",
     "semantic-ui-react": "^0.82.5",
     "terser": "^3.16.1",
@@ -34,7 +34,7 @@
     "mini-css-extract-plugin": "^0.5.0",
     "standard": "^10.0.2",
     "style-loader": "^0.23.1",
-    "webpack": "^4.29.3",
+    "webpack": "^4.41.2",
     "webpack-cli": "^3.2.3",
     "webpack-merge": "^4.2.1"
   },
